Create new files for rolling primitives and add test
